+++ flake.nix
@@ -20,7 +20,7 @@
 
         see = pkgs.rustPlatform.buildRustPackage {
           pname = "see";
-          version = "0.9.0";
+          version = "0.9.1";
 
           src = ./.;
           cargoLock.lockFile = ./Cargo.lock;